:book: Emphasize unicorn/prefer-add-event-listener
I continue to stumble on tutorials and examples that use onclick events in the HTML instead of eventListeners in JavaScript. eventListeners should be considered a best-practice, but almost every time I see examples for beginners, they've never even heard of eventListeners.

on-hooks are more accessible to newcomers. They don't need to get a reference to an element and for basic use-cases they are often shorter than the eventListener alternative, but they're still a bad practice.

I'm adding a star next to the unicorn/prefer-add-event-listener to emphasize its continued importance.